requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
alpaca-py==0.33.1
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
